third_party/blink/web_tests/platform/generic/http/tests/websocket/permessage-deflate-unsolicited-negotiation-response-expected.txt

Test whether WebSocket ignores unsolicited negotiation response.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onopen
PASS ws.extensions is "permessage-deflate; server_max_window_bits=15; server_no_context_takeover"
PASS event.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
